refactor: separate activityItem from its father component
