+++ apps/admin/src/features/dashboard/components/ActivityCard/ActivityCard.tsx
@@ -1,9 +1,9 @@
 import { Divider, Flex, Select } from "antd";
 // @ts-ignore
 import { useStyles } from "./ActivityCardStyles";
-import { Cards, ChatText, Star } from "@phosphor-icons/react";
-import { getRelativeTime } from "@utils";
-import { JSX, useState } from "react";
+
+import { useState } from "react";
+import { ActivityItem } from "./ActivityItem";
 
 
 const items: ActivityItemProps[] = [
@@ -79,45 +79,6 @@ export const ActivityCard = () => {
       </Flex>
     </Flex>
 
-  );
-};
-
-
-const ActivityItem = ({ name, type, content, date }: ActivityItemProps) => {
-
-  const styles = useStyles().styles;
-  const typeTextMap: Record<ActivityItemProps["type"], string> = {
-    comment: "commented on your lecture",
-    rating: "gave a 5 star rating on your course",
-    purchase: "purchased your course",
-  };
-
-  const iconMap: Record<ActivityItemProps["type"], JSX.Element> = {
-    comment: <ChatText size={20} weight="fill" color="white" />,
-    rating: <Star size={20} weight="fill" color="white" />,
-    purchase: <Cards size={20} weight="fill" color="white" />,
-  };
-
-  return (
-    <Flex style={{ padding: "12px 20px" }} gap={"middle"}>
-      <Flex vertical={true} justify={"start"} align={"center"}>
-        <Flex align={"center"} justify={"center"} className={styles.iconCircle}>
-          {iconMap[type]}
-        </Flex>
-      </Flex>
-      <Flex vertical={true} justify={"space-between"} align={"start"} gap={"small"}>
-        <p style={{ margin: 0 }}>
-          <span style={{ fontWeight: "bold" }}>
-            {name}
-          </span>{" "}
-          <span style={{ color: "#888" }}>
-            {typeTextMap[type]}
-          </span>{" "}
-          “{content}”
-        </p>
-        <p style={{ margin: 0, color: "#888" }}>{getRelativeTime(date)}</p>
-      </Flex>
-    </Flex>
   );
 };
